Refeactor LocationModal component
