+++ frontend/src/types/location.ts
@@ -21,3 +21,25 @@ export interface LocationModalProps {
   refreshEvents: () => void;
   mode: 'edit' | 'create';
 }
+
+export interface LocationFormProps {
+  address: string;
+  setAddress: (value: string) => void;
+  editedLocation: any;
+  setEditedLocation: (value: any) => void;
+  handleSearchLocation: () => void;
+}
+
+export interface LocationButtonsProps {
+  handleSaveLocation: () => void;
+  handleDeleteLocation: () => void;
+  onClose: () => void;
+  mode: string;
+  isSaveDisabled: boolean;
+}
+
+export interface LocationMapProps {
+  latitude: number;
+  longitude: number;
+  setEditedLocation: React.Dispatch<React.SetStateAction<Location>>;
+}
